Fuzzer: Adjust feature fuzzing frequency (#6305)
We used to fuzz MVP 1/3, all 1/3, and a mixture 1/3, but that gives far too much
priority to the MVP which is increasingly less important. It is also a good idea to
give "all" more priority as that enables more initial content to run (the fuzzer will
discard initial content if it doesn't validate with the features chosen in the current
iteration).

Also (NFC) rename POSSIBLE_FEATURE_OPTS to make the code easier to follow.