Add a WebAssembly build to release (#6351)
Simply build wasm-opt with Emscripten and bundle that up.

Example build:

https://github.com/kripken/binaryen/releases/tag/wasm-build-1

Specifically

binaryen-wasm-build-1-wasm.tar.gz

Only 1.72 MB, as it's just wasm-opt and not any other tool, so it is
much smaller than our other targets. Perhaps we will add more of the
tools later as needed (wasm-metadce, wasm-split, etc.).

Also update the readme regarding which toolchains use us as a library, that I
noticed while editing it to add the release platforms.