[NFC] Add the type of the Expression when eliding it (#6362)
In some cases we don't print an Expression in full if it is unreachable, so
we print something instead as a placeholder. This happens in unreachable
code when the children don't provide enough info to print the parent (e.g.
a StructGet with an unreachable reference doesn't know what struct type
to use).

This PR prints out the name of the Expression type of such things, which
can help debugging sometimes.