Remove the typed variant of `Select::finalize`
It was not correct because it did not check for unreachability in the select's
children. Rather than fix it, which would have required largely duplicating the
untyped version of `Select::finalize`, simply remove it. This simplifies all the
uses of the old API at the cost of forcing them to compute a LUB, which seems
acceptable.