Fix printing of bulk array ops (#6387)
When the bulk array ops had unreachable or null array types, they were replaced
with blocks, but not using the correct code that also prints all their children
as dropped followed by an unreachable. This meant that the text output in those
cases did not parse as a valid module.

Fix the bug. A follow-up PR will simplify the code to prevent similar bugs from
occurring in the future.