[Strings] Fix precomputing of StringEq (#6401)
We incorrectly overrode the string operations in the interpreter's subclasses. But
string operations can be implemented in the topmost class there (as they depend on
no module state), so just implement them there, once, in a proper way.

This fixes StringEq by removing its override, and moves the others to the right
place.