Remove "minimal" JS import/export legalization (#6428)
This change removes the "minimal" mode from `LegalizeJSInterface`
which was added in #1883.

The idea behind this change was to avoid legalizing most function except
those we know that JS will be calling.  The idea was that for dynamic
linking we always want the non-legalized version to be shared between
wasm module.  These days we solve this problem in a different way with
the `legalize-js-interface-export-originals` which exports the original
functions alongside the legalized ones.  Emscripten then always
prefers the `$orig` functions when doing dynamic linking.