GUFA: Fix hashing of GlobalInfo's type (#6455)
For a global we store the name and a type, and the type may
be more precise than the global's type in the wasm. As a
result, when hashing, it is not enough to hash only the name, so hash
the type as well.

Also add a random TODO as a comment.