Fix ConstantFieldPropagation signed packed field handling and improve Heap2Local's (#6493)
CFP already had logic for truncating but not for sign-extending, which this
fixes.

Use the new helper function in Heap2Local as well. This changes the model
there from "truncate on set, sign-extend on get" to "truncate or sign-extend
on get". That is both simpler by reusing the same logic as CFP but also more
optimal: the idea to truncate on sets made sense since sets are rarer, but if
we must then sign-extend on gets then we can end up doing more work
overall (as the truncations on sets are not needed if all gets are signed).

Found by #6486