Do not repeat types names in text output
For types that do not have explicit names, we generate index-based names in the
printer. However, we did not previously ensure that the generated types were not
already used as explicit names, so it was possible to print the same name for
multiple types, which is not valid.

Fix the problem by skipping indices that are already used as type names.

Fixes #6492.