[NFC] Fix tests that were polluting the test directory (#6498)
A few tests were creating files with the `%s` prefix, which expands to the test
filename itself, so the test files were being created in the test directory and
were never cleaned up. Another test was creating files called `%a.wast` and
`%b.wast`, but `%a` and `%b` do not expand to anything, so this also polluted
the test directory and caused problems with running other tests and fuzzing.

Use the proper `%t` prefix to create all these files in a temporary location
instead.

Fixes #6429.